requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
google-auth>=2.25.0
google-api-python-client>=2.110.0
python-docx>=1.0.0